+++ packages/components/src/instructionsList/InstructionList.story.js
@@ -13,7 +13,14 @@ export const basic = () => {
     'Invite at least 3 friends',
     'Convince them to use this amazing product',
   ]);
-  const donts = array('donts', ['Paying extra hidden fees for transfers', 'Use bad exchange rate']);
+  const donts = array('donts', [
+    <>
+      Paying extra{' '}
+      <a href="https://transferwise.com/gb/blog/why-we-care-about-hidden-fees">hidden fees</a> for
+      transfers
+    </>,
+    'Use bad exchange rate',
+  ]);
 
   return <InstructionsList dos={dos} donts={donts} />;
 };