+++ packages/components/src/typeahead/Typeahead.js
@@ -359,6 +359,7 @@ export default class Typeahead extends Component {
       autoFocus,
       maxHeight,
       alert,
+      inputAutoComplete,
     } = this.props;
 
     const { errorState, query, selected, optionsShown, keyboardFocusedOptionIndex } = this.state;
@@ -414,6 +415,7 @@ export default class Typeahead extends Component {
               onKeyDown={this.handleOnKeyDown}
               onFocus={this.handleOnFocus}
               onPaste={this.handleOnPaste}
+              autoComplete={inputAutoComplete}
             />
 
             {clearButton && (
@@ -473,6 +475,7 @@ Typeahead.propTypes = {
   onFocus: Types.func,
   chipSeparators: Types.arrayOf(Types.string),
   size: Types.oneOf([Typeahead.Size.MEDIUM, Typeahead.Size.LARGE]),
+  inputAutoComplete: Types.string,
 };
 
 Typeahead.defaultProps = {
@@ -497,4 +500,5 @@ Typeahead.defaultProps = {
   onInputChange: null,
   onFocus: null,
   validateChip: () => true,
+  inputAutoComplete: 'new-password',
 };