+++ packages/components/src/typeahead/Typeahead.story.js
@@ -100,6 +100,7 @@ export const basic = () => {
       addon={<i className="input-group-text icon icon-search" />}
       onBlur={() => {}}
       options={options}
+      inputAutoComplete="off"
     />
   );
 };