+++ packages/docs/liveEditorCode/instructionList.code.js
@@ -4,5 +4,8 @@
     'Invite at least 3 friends',
     'Convince them to use this amazing product',
   ]}
-  donts={['Paying extra hidden fees for transfers', 'Use bad exchange rate']}
+  donts={[
+    <>Paying extra <a href="">hidden fees</a> for transfers</>, 
+    'Use bad exchange rate'
+  ]}
 />;