+++ packages/dynamic-flows/src/jsonSchemaForm/JsonSchemaForm.story.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { action } from '@storybook/addon-actions';
-import { boolean, select } from '@storybook/addon-knobs';
+import { boolean, select, text } from '@storybook/addon-knobs';
 import JsonSchemaForm from './JsonSchemaForm';
 
 import simpleSchema from './schemas/simple.json';
@@ -28,12 +28,9 @@ export const basic = () => {
     string: 'hi',
   };
 
-  const errors = {
-    string: 'Manual error',
-  };
-
   const locale = select('locale', ['en-GB', 'jp-JP'], 'en-GB');
-
+  const stringError = text('error from server', '');
+  const errors = { string: stringError };
   const translations = {};
   const submitted = boolean('submitted', false);